# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Navidrome Music Room Gateway API
  version: 1.1.0-dev
  license:
    name: GPL-3.0-only
    identifier: GPL-3.0-only
  description: |
    Versioned room coordination API for MusicMate. Media bytes never pass through this API;
    clients stream audio, artwork, and lyrics directly from Navidrome with their own credentials.
servers:
  - url: https://music.example.com/music-room/api/v1
security:
  - roomSession: []
...
paths:
...
  /rooms/{room_id}/invites:
    parameters: [{ $ref: '#/components/parameters/RoomID' }]
    get:
      tags: [Invitations]
      summary: List invitation metadata without secrets
      responses:
        '200':
          description: Invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  invites:
                    type: array
                    items: { $ref: '#/components/schemas/Invite' }
    post:
      tags: [Invitations]
      summary: Create a revocable invitation
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string, maxLength: 80 }
                expiresAt: { type: string, format: date-time }
                maxUses: { type: integer, minimum: 1, maximum: 10000, default: 20 }
                singleUse: { type: boolean, default: false }
      responses:
        '201':
          description: Secret is returned once; only its SHA-256 digest is persisted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InviteWithSecret' }
...
  /rooms/{room_id}/members:
    get:
      tags: [Rooms]
      summary: List persistent members with current presence
      parameters: [{ $ref: '#/components/parameters/RoomID' }]
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items: { $ref: '#/components/schemas/Member' }
...
components:
  securitySchemes:
    roomSession:
      type: http
      scheme: bearer
      bearerFormat: opaque-15-minute-room-session
  parameters:
    RoomID:
      name: room_id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Identifier' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        A caller-generated key retained in memory for 15 minutes. Repeating the same method,
        path, body, current authorized user, and key replays the original response and adds
        `Idempotency-Replayed: true`; reusing it for a different request returns 409.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
...
  schemas:
    Identifier:
      type: string
      pattern: '^[A-Za-z0-9_-]{16,128}$'
...
    Member:
      type: object
      required: [username, displayName, role, active, online, joinedAt, lastSeenAt]
      properties:
        roomID: { $ref: '#/components/schemas/Identifier' }
        username: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [owner, member] }
        active: { type: boolean }
        online: { type: boolean }
        joinedAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
    Invite:
      type: object
      required: [inviteID, roomID, label, expiresAt, maxUses, useCount, createdBy, createdAt]
      properties:
        inviteID: { $ref: '#/components/schemas/Identifier' }
        roomID: { $ref: '#/components/schemas/Identifier' }
        label: { type: string }
        expiresAt: { type: string, format: date-time }
        maxUses: { type: integer }
        useCount: { type: integer }
        revokedAt: { type: string, format: date-time }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
    InviteWithSecret:
      allOf:
        - $ref: '#/components/schemas/Invite'
        - type: object
          required: [invite, shareURL, deepLink]
          properties:
            invite: { type: string, readOnly: true }
            shareURL: { type: string, format: uri }
            deepLink: { type: string, format: uri }
...